remove final mixin, this one only affected lifecycle methods so it was a bit bigger of a change
