+++ package.json
@@ -30,7 +30,6 @@
     "react": "0.14.6",
     "react-addons-create-fragment": "0.14.6",
     "react-addons-css-transition-group": "0.14.6",
-    "react-addons-pure-render-mixin": "0.14.6",
     "react-addons-transition-group": "0.14.6",
     "react-addons-update": "0.14.6",
     "react-dom": "0.14.6",
