+++ src/components/Collection/CollectionIntro.jsx
@@ -1,12 +1,28 @@
 var React = require('react');
 
-var CollectionIntro = React.createClass({
-  mixins: [
-    require("../../mixins/MuiThemeMixin.jsx")
-  ],
+var mui = require("material-ui");
+var ThemeManager = require('material-ui/lib/styles/theme-manager');
+var BeehiveTheme = require('../../themes/beehive.jsx');
 
+var CollectionIntro = React.createClass({
   propTypes: {
     collection: React.PropTypes.object.isRequired,
+  },
+
+  childContextTypes: {
+    muiTheme: React.PropTypes.object
+  },
+
+  getChildContext() {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
+  },
+
+  getInitialState: function() {
+    return {
+      muiTheme: ThemeManager.getMuiTheme(BeehiveTheme),
+    };
   },
 
   style: function () {
