+++ src/components/CollectionIntroduction/CollectionIntroduction.jsx
@@ -1,6 +1,8 @@
 'use strict'
 var React = require("react");
 var mui = require('material-ui');
+var ThemeManager = require('material-ui/lib/styles/theme-manager');
+var BeehiveTheme = require('../../themes/beehive.jsx');
 
 var CollectionPageHeader = require('../../layout/CollectionPageHeader.jsx');
 var PageContent = require('../../layout/PageContent.jsx');
@@ -11,10 +13,6 @@ var PageTitleBar = require('../Pages/PageTitleBar.jsx');
 const LoadRemote = require('../../modules/LoadRemote.jsx')
 
 var CollectionIntroduction = React.createClass({
-  mixins: [
-    require("../../mixins/MuiThemeMixin.jsx")
-  ],
-
   propTypes: {
     collection: React.PropTypes.oneOfType([
       React.PropTypes.string,
@@ -22,10 +20,21 @@ var CollectionIntroduction = React.createClass({
     ]),
   },
 
+  childContextTypes: {
+    muiTheme: React.PropTypes.object
+  },
+
+  getChildContext() {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
+  },
+
   getInitialState: function() {
     return {
       collection: {},
       remoteCollectionLoaded: false,
+      muiTheme: ThemeManager.getMuiTheme(BeehiveTheme),
     };
   },
