+++ src/components/Error/ErrorPage.jsx
@@ -1,14 +1,28 @@
 'use strict'
 var React = require('react');
 var mui = require('material-ui');
+var ThemeManager = require('material-ui/lib/styles/theme-manager');
+var BeehiveTheme = require('../themes/beehive.jsx');
 
 var PageHeader = require('../../layout/PageHeader.jsx');
 var PageContent = require('../../layout/PageContent.jsx');
 
 var ErrorPage = React.createClass({
-  mixins: [
-    require("../../mixins/MuiThemeMixin.jsx")
-  ],
+  childContextTypes: {
+    muiTheme: React.PropTypes.object
+  },
+
+  getChildContext() {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
+  },
+
+  getInitialState: function() {
+    return {
+      muiTheme: ThemeManager.getMuiTheme(BeehiveTheme),
+    };
+  },
 
   render: function() {
     var url = window.location.origin;
