+++ src/layout/PreviewLink.jsx
@@ -1,6 +1,8 @@
 'use strict'
 var React = require("react");
 var mui = require('material-ui');
+var ThemeManager = require('material-ui/lib/styles/theme-manager');
+var BeehiveTheme = require('../themes/beehive.jsx');
 
 var HoneycombImage = require("../other/HoneycombImage.jsx");
 var MediaQuery = require('react-responsive');
@@ -8,16 +10,24 @@ import { Link } from 'react-router';
 const CollectionUrl = require('../modules/CollectionUrl.jsx')
 
 var PreviewLink = React.createClass({
-  mixins: [
-    require("../mixins/MuiThemeMixin.jsx")
-  ],
   propTypes: {
     siteObject: React.PropTypes.object,
   },
 
+  childContextTypes: {
+    muiTheme: React.PropTypes.object
+  },
+
+  getChildContext() {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
+  },
+
   getInitialState: function () {
     return {
-      hover: false
+      hover: false,
+      muiTheme: ThemeManager.getMuiTheme(BeehiveTheme),
     };
   },
