+++ src/routes/ErrorPage.jsx
@@ -1,18 +1,32 @@
 'use strict'
 var React = require('react');
 var mui = require('material-ui');
+var ThemeManager = require('material-ui/lib/styles/theme-manager');
+var BeehiveTheme = require('../themes/beehive.jsx');
 
 var PageHeader = require('../layout/PageHeader.jsx');
 var PageContent = require('../layout/PageContent.jsx');
 var PageTitle = require('../modules/PageTitle.js')
 
 var ErrorPage = React.createClass({
-  mixins: [
-    require("../mixins/MuiThemeMixin.jsx")
-  ],
-
   componentWillMount() {
     PageTitle("Page Not Found")
+  },
+
+  childContextTypes: {
+    muiTheme: React.PropTypes.object
+  },
+
+  getChildContext() {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
+  },
+
+  getInitialState: function() {
+    return {
+      muiTheme: ThemeManager.getMuiTheme(BeehiveTheme),
+    };
   },
 
   render: function() {
