+++ src/routes/PrintableMetadata.jsx
@@ -1,6 +1,8 @@
 'use strict'
 var React = require('react');
 var mui = require('material-ui');
+var ThemeManager = require('material-ui/lib/styles/theme-manager');
+var BeehiveTheme = require('../themes/beehive.jsx');
 
 var EventEmitter = require('../middleware/EventEmitter.js');
 var HoneycombURL = require('../modules/HoneycombURL.js');
@@ -11,9 +13,21 @@ var ConfigurationStore = require("../store/ConfigurationStore.js");
 const LoadRemote = require('../modules/LoadRemote.jsx')
 
 var PrintableMetadata = React.createClass({
-  mixins: [
-    require("../mixins/MuiThemeMixin.jsx")
-  ],
+  childContextTypes: {
+    muiTheme: React.PropTypes.object
+  },
+
+  getChildContext() {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
+  },
+
+  getInitialState: function() {
+    return {
+      muiTheme: ThemeManager.getMuiTheme(BeehiveTheme),
+    };
+  },
 
   componentWillMount: function() {
     ConfigurationStore.addChangeListener(this.configurationLoaded);
